[source build] fix dependency problems
`colcon-common-extensions`' fetched a later version of `colcon_python_setup_py`, which in turn, required a later version of `colcon_core` that resulted in a version conflict, refer to https://github.com/kobuki-base/kobuki_documentation/actions/runs/339815780/workflow. This commit takes out the potential for a version conflict and let's pip do it's thing.